# performance: set TCP_NODELAY and increase copy buffer