io::copy default buffer is 8k, we see increased performance with a more precise 16k buffer.

Results:

Throughput Gb/s
buffer nodelay: 7
nodelay: 4.5
neither: 3.5

Latency p50,p90,p99
buffer nodelay: 0.17ms  0.21ms  0.32ms
nodelay:  0.17ms  0.22ms  0.46ms
neither: 0.17ms  0.22ms  0.54ms

("Neither" is before, "buffer nodelay" is this PR)