@howardjohn 2022-11-14
oddly setting it on outbound drops us ~1Gb/s worse. Not sure why. Needs more investigation. Might make sense to merge this as-is for now and continue to tune it moving forward; hopefully soon we will have proper benchmarks as well.