@hzxuzhonghu 2022-11-15
great test, can you share the test script. So anyone want to optimize some thing, it would be possible to reuse your effort