@howardjohn 2022-11-15
LOCAL.md with simple iperf3

On Mon, Nov 14, 2022 at 6:50 PM Zhonghu Xu ***@***.***> wrote:

> great test, can you share the test script. So anyone want to optimize some
> thing, it would be possible to reuse your effort
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/istio/ztunnel/pull/109#issuecomment-1314688121>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEYGXNJB2IMHQDWYZVMNG3WIL267ANCNFSM6AAAAAAR55RB2M>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>